requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
google-generativeai
python-dotenv
mistune
Flask-Mail